+++ modules/apps/client-extension/client-extension-type-api/bnd.bnd
@@ -5,6 +5,7 @@ Export-Package:\
 	com.liferay.client.extension.type,\
 	com.liferay.client.extension.type.configuration,\
 	com.liferay.client.extension.type.deployer,\
+	com.liferay.client.extension.type.facet,\
 	com.liferay.client.extension.type.factory,\
 	com.liferay.client.extension.type.manager,\
 	com.liferay.client.extension.type.util,\
