+++ modules/apps/client-extension/client-extension-type-api/src/main/resources/com/liferay/client/extension/type/facet/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
