+++ modules/apps/client-extension/client-extension-type-impl/src/main/java/com/liferay/client/extension/type/internal/CETGlobalJSImpl.java
@@ -21,8 +21,6 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.UnicodePropertiesBuilder;
 
-import java.util.Properties;
-
 import javax.portlet.PortletRequest;
 
 /**
@@ -48,12 +46,12 @@ public class CETGlobalJSImpl extends BaseCETImpl implements CETGlobalJS {
 
 	public CETGlobalJSImpl(
 		String baseURL, long companyId, String description,
-		String externalReferenceCode, String name, Properties properties,
-		String sourceCodeURL, UnicodeProperties typeSettingsUnicodeProperties) {
+		String externalReferenceCode, String name, String sourceCodeURL,
+		UnicodeProperties typeSettingsUnicodeProperties) {
 
 		super(
 			baseURL, companyId, description, externalReferenceCode, name,
-			properties, sourceCodeURL, typeSettingsUnicodeProperties);
+			sourceCodeURL, typeSettingsUnicodeProperties);
 	}
 
 	public CETGlobalJSImpl(UnicodeProperties typeSettingsUnicodeProperties) {
